+++ host_test/extra_dirs_common.cmake
@@ -9,7 +9,7 @@ get_filename_component(PROJECT_ROOT "${CMAKE_CURRENT_LIST_DIR}/.." ABSOLUTE)
 list(APPEND EXTRA_COMPONENT_DIRS
     "${PROJECT_ROOT}"
     "${PROJECT_ROOT}/host_test/gtest"
-    "${PROJECT_ROOT}/host_test/common"
+    "${PROJECT_ROOT}/host_test/mocks"
     "${PROJECT_ROOT}/external"
     "$ENV{IDF_PATH}/tools/mocks/esp_timer"
     "$ENV{IDF_PATH}/tools/mocks/driver"