+++ host_test/mocks/mock_battery_monitor.hpp
@@ -1,0 +1,4 @@
+// components/battery_monitor/host_test/mocks/mock_battery_monitor.hpp
+#pragma once
+
+#include "mock_i_battery_monitor.hpp"